ir: Move dimension_sort to ir/equations
